docs(Typography): group heading, ingress, muted and info and enhance prose
