+++ designsystem/card/card.mdx
@@ -9,6 +9,7 @@ import * as stories from './card.stories';
 
 # Card og Group
 - Bruk klassen `card` på typisk `<div>`, `<a>` eller `<button>`
+- Bruk [`info`](?path=/story/designsystem-typography--info-story) rundt ledetekst+verdi, ikon+verdi eller verdi for enkelt, responsivt oppsett
 <Canvas of={stories.Default} />
 
 ## Interaktiv
@@ -29,6 +30,7 @@ Frykt ikke dog! Du kan kontakt oss på [#designsystem](https://mattilsynet-hq.sl
 
 ## Gruppering
 - Bruk `group` på typisk en `grid` eller `flex` for å gruppere flere `card`
+- Kan også benyttes rundt annet innhold for å visuelt vise gruppering
 <Canvas of={stories.GroupStory} />
 
 <CssVariables component="card" />
