+++ designsystem/errorsummary/errorsummary.module.css
@@ -30,8 +30,3 @@
 		}
 	}
 }
-
-/* TODO: TMP workaround to avoid focus in storybook */
-[data-story-block="true"] .errorsummary {
-	animation: none;
-}
