--- designsystem/heading/heading.mdx
@@ -1,17 +1,0 @@
-import { Meta, Canvas } from '@storybook/blocks';
-import { CssVariables } from '../../.storybook/blocks';
-import * as stories from './heading.stories';
-
-<Meta of={stories} />
-
-# Heading
-- Bruk klassen `heading` på typisk `<h1>` til `<h6>`
-- Bruk `data-justify="center"` for å midtstille tekst
-<Canvas of={stories.Default} />
-
-## Størrelser
-- Bruk `data-size="2xs | xs | sm | md | lg | xl | 2xl"`
-- **Merk:** `heading` endrer kun størrelse - ikke overskriftsnivå.<br />Benytt [korrekt HTML h-tag for god universell utforming &rarr;](https://www.uutilsynet.no/veiledning/246-overskrifter-og-ledetekster/1253)
-<Canvas of={stories.Sizes} />
-
-<CssVariables component="heading" />
