--- designsystem/heading/heading.module.css
@@ -1,11 +1,0 @@
-@layer mt.design {
-	.heading {
-		composes: ds-heading from "@digdir/designsystemet-css";
-
-		scroll-margin-block: 4em; /* Ensure scrolling to headings still have some comfortable spacing above */
-		text-wrap: pretty;
-	}
-	.heading[data-justify="center"] {
-		text-align: center;
-	}
-}
