+++ designsystem/overview.mdx
@@ -15,7 +15,6 @@ import * as Divider from './divider/divider.stories.tsx';
 import * as Errorsummary from './errorsummary/errorsummary.stories.tsx';
 import * as Field from './field/field.stories.tsx';
 import * as Fieldset from './fieldset/fieldset.stories.tsx';
-import * as Heading from './heading/heading.stories.tsx';
 import * as Helptext from './helptext/helptext.stories.tsx';
 import * as Input from './input/input.stories.tsx';
 import * as Layout from './layout/layout.stories.tsx';
@@ -29,13 +28,13 @@ import * as Spinner from './spinner/spinner.stories.tsx';
 import * as Table from './table/table.stories.tsx';
 import * as Tabs from './tabs/tabs.stories.tsx';
 import * as Tag from './tag/tag.stories.tsx';
-import * as Text from './text/text.stories.tsx';
 import * as Togglegroup from './togglegroup/togglegroup.stories.tsx';
 import * as Tooltip from './tooltip/tooltip.stories.tsx';
+import * as Typography from './typography/typography.stories.tsx';
 import * as Validation from './validation/validation.stories.tsx';
 
 <Meta title="Designsystem/Komponenter" />
 
 # Komponenter
 
-<Overview items={[Alert, App, Avatar, Badge, Breadcrumbs, Button, Card, Chip, Details, Dialog, Divider, Errorsummary, Field, Fieldset, Heading, Helptext, Input, Layout, Link, Logo, Pagination, Popover, Progress, Skeleton, Spinner, Table, Tabs, Tag, Text, Togglegroup, Tooltip, Validation]} />
+<Overview items={[Alert, App, Avatar, Badge, Breadcrumbs, Button, Card, Chip, Details, Dialog, Divider, Errorsummary, Field, Fieldset, Helptext, Input, Layout, Link, Logo, Pagination, Popover, Progress, Skeleton, Spinner, Table, Tabs, Tag, Togglegroup, Tooltip, Typography, Validation]} />
