+++ designsystem/react.tsx
@@ -12,7 +12,7 @@ export * from "./divider/divider";
 export * from "./errorsummary/errorsummary";
 export * from "./field/field";
 export * from "./fieldset/fieldset";
-export * from "./heading/heading";
+export * from "./typography/typography";
 export * from "./helptext/helptext";
 export * from "./input/input";
 export * from "./layout/layout";
@@ -26,6 +26,5 @@ export * from "./spinner/spinner";
 export * from "./table/table";
 export * from "./tabs/tabs";
 export * from "./tag/tag";
-export * from "./text/text";
 export * from "./togglegroup/togglegroup";
 export * from "./validation/validation";
