+++ designsystem/table/table.mdx
@@ -1,6 +1,6 @@
 import { Meta, Canvas } from '@storybook/blocks';
 import { CssVariables } from '../../.storybook/blocks';
-import styles from "../styles.module.css";
+import { Buttom } from "../react";
 import * as stories from './table.stories';
 
 <Meta of={stories} />
@@ -64,28 +64,28 @@ import * as stories from './table.stories';
 - Plasser tabellen inni en `<figure>` for å få scroll på mobil
 - Bruk `data-nowrap` hvor du vil på tabellen, rader eller celler for å hindre linjeskift
 <div className="sbdocs-canvas-thumbnail">
-  <a className={styles.button} data-arrow data-variant="primary" href="?path=/story/designsystem-table--mobile-scroll">Se scroll tabell i egen visning</a>
+  <Button data-arrow data-variant="primary" href="?path=/story/designsystem-table--mobile-scroll">Se scroll tabell i egen visning</Button>
   <Canvas className="sbdocs-canvas--max-height" of={stories.MobileDivided} />
 </div>
 
 ## Mobile: Divided
 - Legg `data-mobile="divided"` på `<table>` for todelt visning opp til `768px`
 <div className="sbdocs-canvas-thumbnail">
-  <a className={styles.button} data-arrow data-variant="primary" href="?path=/story/designsystem-table--mobile-divided">Se divided tabell i egen visning</a>
+  <Button data-arrow data-variant="primary" href="?path=/story/designsystem-table--mobile-divided">Se divided tabell i egen visning</Button>
   <Canvas className="sbdocs-canvas--max-height" of={stories.MobileDivided} />
 </div>
 
 ## Mobile: Spaced
 - Legg `data-mobile="spaced"` på `<table>` for todelt visning opp til `768px`
 <div className="sbdocs-canvas-thumbnail">
-  <a className={styles.button} data-arrow data-variant="primary" href="?path=/story/designsystem-table--mobile-spaced">Se spaced tabell i egen visning</a>
+  <Button data-arrow data-variant="primary" href="?path=/story/designsystem-table--mobile-spaced">Se spaced tabell i egen visning</Button>
   <Canvas className="sbdocs-canvas--max-height" of={stories.MobileSpaced} />
 </div>
 
 ## Mobile: Stacked
 - Legg `data-mobile="stacked"` på `<table>` for stablet visning opp til `768px`
 <div className="sbdocs-canvas-thumbnail">
-  <a className={styles.button} data-arrow data-variant="primary" href="?path=/story/designsystem-table--mobile-stacked">Se stacked tabell i egen visning</a>
+  <Button data-arrow data-variant="primary" href="?path=/story/designsystem-table--mobile-stacked">Se stacked tabell i egen visning</Button>
   <Canvas className="sbdocs-canvas--max-height" of={stories.MobileStacked} />
 </div>
