+++ designsystem/togglegroup/togglegroup.stories.tsx
@@ -1,3 +1,8 @@
+import {
+	TextAlignCenter,
+	TextAlignLeft,
+	TextAlignRight,
+} from "@phosphor-icons/react";
 import type { Meta, StoryObj } from "@storybook/react";
 import { useState } from "react";
 import { Togglegroup } from "../react";
@@ -12,24 +17,45 @@ type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
 	render: () => (
-		<fieldset className={styles.togglegroup} aria-label="Mappe">
-			<label>
-				<input type="radio" name="my-toggles" value="inboks" />
-				Inboks
-			</label>
-			<label>
-				<input type="radio" name="my-toggles" value="utkast" defaultChecked />
-				Utkast
-			</label>
-			<label>
-				<input type="radio" name="my-toggles" value="arkiv" />
-				Arkiv
-			</label>
-			<label>
-				<input type="radio" name="my-toggles" value="sendt" />
-				Sendt
-			</label>
-		</fieldset>
+		<div className={styles.grid}>
+			<fieldset className={styles.togglegroup} aria-label="Mappe">
+				<label>
+					<input type="radio" name="my-toggles" value="inboks" />
+					Inboks
+				</label>
+				<label>
+					<input type="radio" name="my-toggles" value="utkast" defaultChecked />
+					Utkast
+				</label>
+				<label>
+					<input type="radio" name="my-toggles" value="arkiv" />
+					Arkiv
+				</label>
+				<label>
+					<input type="radio" name="my-toggles" value="sendt" />
+					Sendt
+				</label>
+			</fieldset>
+			<fieldset className={styles.togglegroup} aria-label="Mappe">
+				<label>
+					<input type="radio" name="my-icon-toggle" value="left" />
+					<TextAlignLeft />
+				</label>
+				<label>
+					<input
+						type="radio"
+						name="my-icon-toggle"
+						value="center"
+						defaultChecked
+					/>
+					<TextAlignCenter />
+				</label>
+				<label>
+					<input type="radio" name="my-icon-toggle" value="right" />
+					<TextAlignRight />
+				</label>
+			</fieldset>
+		</div>
 	),
 };
