+++ designsystem/tokens.mdx
@@ -1,8 +1,7 @@
 import { Meta } from '@storybook/blocks';
 import { JumpTo } from '../.storybook/blocks';
 import { Colors, Radius, Sizes } from './tokens';
-import styles from './styles.module.css';
-
+import { Divider } from './react';
 
 <Meta title="Designsystem/Tokens" />
 # Tokens
@@ -16,36 +15,35 @@ Vi benytter semantiske farge-tokens, for å knytte farger til meningsfulle rolle
 Dette gir skalerbarhet, enklere vedlikehold, bedre tilgjengelighet og fleksibilitet.
 Farge-tokens er navngitt `--mtds-color-NAVN-BRUKSOMRÅDE-VARIANT`.
 
-<div className={styles.alert} data-color="success">
-**Tips:**
-- Bruk tokens *uten* spesifisert farge-palett - f.eks. `--mtds-color-BRUKSOMRÅDE-VARIANT`
-- Sett heller fargepalett med attributten `data-color="NAVN"` - da arves paletten
-- Bruk kun `text-*` og `base-contrast-*` på tekst/ikoner - da er kontrastkrav ivaretatt
-- Unngå å blande farger fra ulike paletter for å ivarta kontrastkrav og dark mode
-- Trenger du et mørkt område i light mode? Bruk `data-color="inverted"`
-</div>
+> ✅ **Tips:**
+> - Bruk tokens *uten* spesifisert farge-palett - f.eks. `--mtds-color-BRUKSOMRÅDE-VARIANT`
+> - Sett heller fargepalett med attributten `data-color="NAVN"` - da arves paletten
+> - Bruk kun `text-*` og `base-contrast-*` på tekst/ikoner - da er kontrastkrav ivaretatt
+> - Unngå å blande farger fra ulike paletter for å ivarta kontrastkrav og dark mode
+> - Trenger du et mørkt område i light mode? Bruk `data-color="inverted"`
+
 <br />
 <Colors />
 
-<hr className={styles.divider} style={{ marginBlock: 60 }} />
+<Divider data-gap={15} />
 
 ## Dark mode
 
 Fargemodus kan settes med `data-color-scheme="auto | light | dark"` på `<html>`, hvor `auto` er default og følger brukerens systeminnstillinger.
 
-<hr className={styles.divider} style={{ marginBlock: 60 }} />
+<Divider data-gap={15} />
 
 ## Størrelser
 Størrelser tar utgangspunktet i til en hver tid gjeldene `font-size`, der utgangspunktet er `md` (medium).
 Det betyr at du ikke alltid kjenner den eksakte pikselverdien, men i stedet kan fokusere på forholdet mellom elementene.
 <Sizes />
 
-<hr className={styles.divider} style={{ marginBlock: 60 }} />
+<Divider data-gap={15} />
 
 ## Radius
 <Radius />
 
-<hr className={styles.divider} style={{ marginBlock: 60 }} />
+<Divider data-gap={15} />
 
 ## Tailwind
 Ved å laste `@import '@mattilsynet/design/tailwind.css';` får du automatis designsystemets `--color`, `--spacing`, `--border-radius` og `--font-weight` i Tailwind v4 oppsett.
