+++ designsystem/typography/typography.module.css
@@ -1,0 +1,83 @@
+@layer mt.design {
+	.heading {
+		composes: ds-heading from "@digdir/designsystemet-css";
+
+		scroll-margin-block: 4em; /* Ensure scrolling to headings still have some comfortable spacing above */
+		text-wrap: pretty;
+	}
+	.heading[data-justify="center"] {
+		text-align: center;
+	}
+
+	.ingress {
+		font-size: 1.333333em;
+	}
+	.muted,
+	small {
+		font-size: max(.888888em, .75rem);
+	}
+
+	/**
+	 * Info
+	 */
+	.info {
+		--mtdsc-info-gap: var(--ds-size-2);
+		--mtdsc-info-icon-size: 1em;
+		--mtdsc-info-line-height: var(--ds-line-height-sm);
+		--x: calc(var(--mtdsc-info-icon-size) + var(--mtdsc-info-gap));
+
+		line-height: var(--mtdsc-info-line-height);
+		overflow: clip;
+		text-overflow: ellipsis;
+	}
+	.info strong {
+		display: block;
+	}
+	.info:has(> :is(svg, i)) {
+		padding-left: var(--x);
+	}
+	.info > :is(svg, i) {
+		height: calc(var(--mtdsc-info-line-height) * 1em); /* Fallback if not supporting lh unit */
+		height: 1lh; /* But use 1lh if available to align with line-height */
+		margin-left: calc(var(--x) * -1);
+		position: absolute;
+		text-align: center;
+		width: var(--mtdsc-info-icon-size);
+	}
+
+	/**
+	 * Prose
+	 */
+	.prose {
+		--mtdsc-prose-max-width: 65ch;
+	}
+	.prose > * {
+		margin-block: 1em;
+	}
+	.prose > :first-child {
+		margin-top: 0;
+	}
+	.prose > :last-child {
+		margin-bottom: 0;
+	}
+	.prose > :where(p, ul, ol) {
+		max-width: var(--mtdsc-prose-max-width);
+	}
+	.prose > :where(ul, ol):not([class]) {
+		padding-left: 1em;
+	}
+	.prose > :where(ul, ol):not([class]) > li:not(:first-child) {
+		margin-top: var(--ds-size-2);
+	}
+	.prose > :where([class*="heading"]) {
+		margin-block: 1.5em .5rem; /* Needs to be REM an not em to always align with size tokens, as headings have different sizes */
+		max-width: var(--mtdsc-prose-max-width);
+	}
+	.prose > :where([class*="heading"]) + :where(p, ul, ol, [class*="heading"]) {
+		margin-top: 0;
+	}
+	.prose > :where(button) {
+		display: flex; /* Swap to flex instead of inline-flex to make margins collapse */
+		width: fit-content;
+	}
+}
