+++ identitet/index.mdx
@@ -1,19 +1,19 @@
 import { Meta } from '@storybook/blocks';
-import styles from '../designsystem/styles.module.css';
+import { Button, Flex, Grid, Prose } from '../designsystem/react';
 
 <Meta title="Introduksjon" />
 
-<div style={{ minHeight: 'calc(100vh - 200px)', display: 'grid', placeContent: 'center' }}>
-<div>
+<Grid style={{ minHeight: 'calc(100vh - 200px)', placeContent: 'center' }}>
+<Prose>
 <img src="/illustrations/person-animal-nature-1.svg" alt="" style={{ float: 'right', marginLeft: '1em' }} />
 # Mattilsynet Design
 
 > Mattilsynets visuelle identitet og designsystem hjelper deg å lage gode tjenester for våre brukere.
 
-<div className={styles.flex}>
-<a className={styles.button} data-variant="secondary" href="?path=/docs/identitet-kompasset-og-roller--docs">Identitet</a>
-<a className={styles.button} data-variant="secondary" href="?path=/docs/designsystem-introduksjon--docs">Designsystem</a>
-<a className={styles.button} data-variant="secondary" href="?path=/docs/profilering-maler--docs">Profilering</a>
-</div>
-</div>
-</div>
+<Flex>
+<Button data-variant="secondary" href="?path=/docs/identitet-kompasset-og-roller--docs">Identitet</Button>
+<Button data-variant="secondary" href="?path=/docs/designsystem-introduksjon--docs">Designsystem</Button>
+<Button data-variant="secondary" href="?path=/docs/profilering-maler--docs">Profilering</Button>
+</Flex>
+</Prose>
+</Grid>
