+++ identitet/kompasset.mdx
@@ -1,7 +1,6 @@
 import { Meta } from '@storybook/blocks';
 import kompasset from './kompasset.jpg';
 import { FilePdf } from '@phosphor-icons/react'
-import styles from '../designsystem/styles.module.css';
 import { Button, Details, Flex, Grid } from '../designsystem/react';
 
 <Meta title="Identitet/Kompasset og roller" />
