+++ identitet/typografi.mdx
@@ -1,7 +1,6 @@
 import { Meta } from '@storybook/blocks';
 import { FileArrowDown } from "@phosphor-icons/react";
 import { Alert, Button, Flex, Details } from '../designsystem/react';
-import styles from '../designsystem/styles.module.css';
 
 <Meta title="Identitet/Typografi" />
